requirements.txt
alembic==1.9.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.46
urllib3==2.2.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.18.1
